web/requirements.txt
Flask==0.12.2
gunicorn==19.3.0
RPi.GPIO==0.6.3